App/Moonshine/Layouts and databases patch

+++ database/migrations/2025_03_24_193702_rename_sale_price_to_products_table.php
@@ -12,14 +12,21 @@ return new class extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->decimal('sale_price', 8, 2)->nullable()->after('price');
+            if (!Schema::hasColumn('products', 'sale_price')) {
+                $table->decimal('sale_price', 8, 2)->after('price')->nullable();
+            }
         });
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('sale_price');
+            if (Schema::hasColumn('products', 'sale_price')) {
+                $table->dropColumn('sale_price');
+            }
         });
     }
